+++ packages/core/framework7.esm.js
@@ -1,5 +1,5 @@
 /**
- * Framework7 3.0.0-beta.18
+ * Framework7 3.0.0-beta.19
  * Full featured mobile HTML framework for building iOS & Android apps
  * http://framework7.io/
  *