+++ packages/core/modules/router/process-route-queue.js
@@ -39,11 +39,11 @@ function processQueue(router, routerQueue, routeQueue, to, from, resolve, reject
 export default function (to, from, resolve, reject) {
   const router = this;
   function enterNextRoute() {
-    if (to && to.route && (router.params.beforeEnter || to.route.beforeEnter)) {
+    if (to && to.route && (router.params.routesBeforeEnter || to.route.beforeEnter)) {
       router.allowPageChange = false;
       processQueue(
         router,
-        router.params.beforeEnter,
+        router.params.routesBeforeEnter,
         to.route.beforeEnter,
         to,
         from,
@@ -60,11 +60,11 @@ export default function (to, from, resolve, reject) {
     }
   }
   function leaveCurrentRoute() {
-    if (from && from.route && (router.params.beforeLeave || from.route.beforeLeave)) {
+    if (from && from.route && (router.params.routesBeforeLeave || from.route.beforeLeave)) {
       router.allowPageChange = false;
       processQueue(
         router,
-        router.params.beforeLeave,
+        router.params.routesBeforeLeave,
         from.route.beforeLeave,
         to,
         from,