+++ packages/react/framework7-react.js
@@ -1,5 +1,5 @@
 /**
- * Framework7 React 3.0.0-beta.18
+ * Framework7 React 3.0.0-beta.19
  * Build full featured iOS & Android apps using Framework7 & React
  * http://framework7.io/react/
  *
@@ -10626,7 +10626,7 @@
   };
 
   /**
-   * Framework7 React 3.0.0-beta.18
+   * Framework7 React 3.0.0-beta.19
    * Build full featured iOS & Android apps using Framework7 & React
    * http://framework7.io/react/
    *